requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
RPi.GPIO==0.7.1
Adafruit-DHT==1.4.0
Werkzeug==2.3.7
python-dotenv==1.0.0
schedule==1.2.0

